biblio-hub/server/requirements.txt
wheel==0.30.0
Flask_Cors==3.0.8
pip==19.3.1
Flask==1.1.1
setuptools==39.0.1
pyjwt
redis
flask-swagger-ui==3.20.9
